// frontend/src/api/generated/types.gen.ts
export type AccessListResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    items: Array<AccessResponse>;
};

export type AccessResponse = {
    applicationCode: string;
    canRead: boolean;
    canWrite: boolean;
    createdAt: string;
    id: string;
    roleCode: string;
};
// ...
export type ConfigListResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    items: Array<ConfigResponse>;
};

export type ConfigResponse = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    applicationCode: string;
    clientId?: string;
    createdAt: string;
    description?: string;
    id: string;
    property: string;
    scope: string;
    section: string;
    updatedAt: string;
    value: string;
    valueType: string;
};